First play around with HAL for tiddlyweb
This still needs quite a lot of work as I'm still trying to
figure out how things ought to represented. What's come out
of the work so far is a concept of a HalDocument object, a Links
object, and Link object which attempt to encapsulate the HAL
structure before ever going to JSON. Initially I assumed I
would just build links but having the classes encapsulates
some of the logic. I suspect things will change quite a bit
as it doesn't yet feel natural.

In terms of HAL, I find the syntax around embedding to be a
bit awkward. For example with a list of bags we end up with:

  {"_links": {"self": {"href": "/bags"}},
   "_embedded": {
       "bag": [
         {"_links": {"self": {"href": "/bags/bag0"}}, "name": "bag0"},
         {"_links": {"self": {"href": "/bags/bag1"}}, "name": "bag1"},
         {"_links": {"self": {"href": "/bags/bag2"}}, "name": "bag2"},
         {"_links": {"self": {"href": "/bags/bag3"}}, "name": "bag3"},
         {"_links": {"self": {"href": "/bags/bag4"}}, "name": "bag4"}
       ]
   }
  }

which at first glance seems rough to me, but I'm guessing that the goal
is to be able to colletions of different stuff embedded, not just the
one kind of thing.

See the README.md for other questions.