Add support for single tiddlers
And adjust for singl revisions and revision collections.

At this point it is now getting pretty clear that we need a
code driven source for link rels, because that's really all
this code is: figuring out the rels. Clearly you can make a ton
of them, but is that truly needed?

Other issues:

* _tiddler_dict is incomplete, compared with the json rep
* missing rels for things like tiddlyweb:bag_edit