Get rid of the tiddler_edit link rel
This is not needed: if PUT JSON to tiddlyweb:tiddler you'll
get one. In fact if you PUT a HAL+JSON rep to tiddlyweb:tiddler
you'll get a tiddler (because such a rep would not be _embedded, and
weird fields would be ignored). However it would have to be PUT
as application/json not, application/hal+json.

Not sure which way that ought to go. Could subclass the json rep
and let the superclass handle all the as_* ?