add test/test_ops
This lead to extensive changes throughout (and in the hal module)

* present full uris, so we don't ever have to calculate
* add tiddlyweb:bag, tiddlweb:recipe and tiddlyweb:tiddler links
  to the bags, recipes, and tiddlers collections, respectively

I'm still not entirely convinced. Now instead of having to remember
uris, I have to remember link rels. That's fine if I'm assuming
that the links will change, but damn dood, these links are cool, so
why would they?